move the concatenateArrays function as it may be generally useful with VolumeMaker (in vole-app for example)
